// src/app/dto/clientes/cliente-dto.ts
export class ClienteDto {
    nome: string;
    cnpj_Cpf: string;
    rg: string;
    cep: string;
    endereco: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    uf: string;
    telefoneResidencial: string;
    celular: string;
    telComercial: string;
    ramal: string;
    observacao_Filiacao: string;
    email: string;
    emailXml: string;
    produtorRural: string;
    DadosCliente:any;
}